Factorize filename logics between local and gcloud
