+++ src/kinto_attachment/storage/__init__.py
@@ -1,7 +1,9 @@
 import os
 import re
 import unicodedata
+import urllib
 import uuid
+from datetime import datetime
 
 from pyramid import exceptions as pyramid_exceptions
 from zope.interface import Attribute, Interface
@@ -19,7 +21,15 @@ class IFileStorage(Interface):
     def exists(filename):
         """Return ``True`` if *filename* exists in the store."""
 
-    def save(fs, folder=None, randomize=False, filename_pattern=None, record_id="", replace=False, headers=None):
+    def save(
+        fs,
+        folder=None,
+        randomize=False,
+        filename_pattern=None,
+        record_id="",
+        replace=False,
+        headers=None,
+    ):
         """Persist *fs* (a ``cgi.FieldStorage``-like object) and return the
         stored relative filename.
 
@@ -34,6 +44,29 @@ class IFileStorage(Interface):
 
     def delete(filename):
         """Remove *filename* from the store."""
+
+
+class FileStorage:
+    base_url = ""
+
+    def url(self, filename):
+        return urllib.parse.urljoin(self.base_url, filename)
+
+    def save(self, fs, *args, **kwargs):
+        return self.save_file(fs.file, fs.filename, *args, **kwargs)
+
+    @staticmethod
+    def _prepare_filename(filename, randomize=False, filename_pattern=None, record_id=""):
+        filename = secure_filename(os.path.basename(filename))
+        if randomize:
+            filename = random_filename(filename)
+        if filename_pattern:
+            filename = filename_pattern.format(
+                datetime=datetime.now().strftime("%Y%m%d%H%M%S"),
+                rid=record_id,
+                filename=filename,
+            )
+        return filename
 
 
 def register_file_storage_impl(config, impl):
