+++ tests/storage/test_gcloud.py
@@ -123,7 +123,7 @@ class TestGoogleCloudStorage(unittest.TestCase):
         self.assertTrue(self.storage.exists(name))
 
     def test_save_file_filename_pattern_datetime(self):
-        with mock.patch("kinto_attachment.storage.gcloud.datetime") as mock_datetime:
+        with mock.patch("kinto_attachment.storage.datetime") as mock_datetime:
             mock_datetime.now.return_value = datetime.datetime(2024, 6, 1, 1, 1, 1)
             name = self.storage.save_file(
                 io.BytesIO(b"data"), "file.txt", filename_pattern="{datetime}-{filename}"
@@ -132,7 +132,7 @@ class TestGoogleCloudStorage(unittest.TestCase):
         self.assertTrue(self.storage.exists(name))
 
     def test_save_file_filename_pattern_with_rid(self):
-        with mock.patch("kinto_attachment.storage.gcloud.datetime") as mock_datetime:
+        with mock.patch("kinto_attachment.storage.datetime") as mock_datetime:
             mock_datetime.now.return_value = datetime.datetime(2024, 6, 1, 1, 1, 1)
             name = self.storage.save_file(
                 io.BytesIO(b"data"),
