+++ tests/storage/test_local.py
@@ -91,7 +91,7 @@ class TestLocalFileStorage(unittest.TestCase):
         self.assertTrue(self.storage.exists(name))
 
     def test_save_file_filename_pattern_datetime(self):
-        with mock.patch("kinto_attachment.storage.local.datetime") as mock_datetime:
+        with mock.patch("kinto_attachment.storage.datetime") as mock_datetime:
             mock_datetime.now.return_value = datetime.datetime(2024, 6, 1, 1, 1, 1)
             name = self.storage.save_file(
                 io.BytesIO(b"data"), "file.txt", filename_pattern="{datetime}-{filename}"
@@ -100,7 +100,7 @@ class TestLocalFileStorage(unittest.TestCase):
         self.assertTrue(self.storage.exists(name))
 
     def test_save_file_filename_pattern_with_rid(self):
-        with mock.patch("kinto_attachment.storage.local.datetime") as mock_datetime:
+        with mock.patch("kinto_attachment.storage.datetime") as mock_datetime:
             mock_datetime.now.return_value = datetime.datetime(2024, 6, 1, 1, 1, 1)
             name = self.storage.save_file(
                 io.BytesIO(b"data"),
